Update the target.f for cryo17 loop1 and loop2
Add the proper materials for loop1 (targ%can=3)
and loop2 (targ%can=4) 2017 cryotargets.

Add proper calculation of target length in
the cryotarget with the beer can topped with
a half-sphere.

Set the SHMS material for the scattering chamber
window, air and HB entrance correctly.

Set the HMS material for scattering chamber,
air and kevlar/mylar window.